Handle websocket errors in KubeExec
Prior to this commit, the KubeExec class didn't verify that
there were no errors in the stream and returned the messages to
the caller. Then, the caller (in this case 'InjectClientCert')
parsed the messages and verified there were no errors. Instead,
it is better to handle the errors in the KubeExec class and raise
an error if one occurred.

In addition, we no longer need to have the class implemented in
the bottom of the file as we don't use the `execute` method anymore.